Move relevant concrete user identity implementations to infrastructure layer, and move interfaces to Common namespace
