+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/UserIdentity/UserIdentityExtensions.cs
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MeteringPoints.Application.UserIdentity
+namespace Energinet.DataHub.MeteringPoints.Infrastructure.UserIdentity
 {
     public static class UserIdentityExtensions
     {
-        public static string AsString(this UserIdentity userIdentity)
+        public static string AsString(this Application.Common.UserIdentity.UserIdentity userIdentity)
         {
             return System.Text.Json.JsonSerializer.Serialize(userIdentity);
         }
